OBS Bug Fixes
Popups now loop properly, hopefully.
OBS Service when failed to connect will not "connect" on the front end.
Popups now use a monitor name instead of index, however fundamentally works the same via an index.
Popups are handled via a queue, to prevent overlap just in case, but it seems this is no longer needed, but I'll be keeping just in case.